pyproject.toml
[tool.poetry]
name = "lacrei-saude-api"
version = "1.0.0"
description = "API RESTful de Gerenciamento de Consultas Médicas - Lacrei Saúde"
authors = ["Carlos Alberto Furtado"]
readme = "README.md"
packages = [{include = "core"}, {include = "apps"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.6"
django-filter = "^24.3"
drf-spectacular = "^0.28"
psycopg2-binary = "^2.9"
python-decouple = "^3.8"
gunicorn = "^23.0"
bleach = "^6.2"
whitenoise = "^6.8"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1"
black = "^24.10"
isort = "^5.13"
coverage = "^7.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
known_django = ["django", "rest_framework"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
